修改权限
修改权限

+++ web/src/main/java/com/staging/controller/RoleController.java
@@ -7,10 +7,13 @@ import com.staging.common.PagerLayui;
 import com.staging.common.ServerResponse;
 import com.staging.common.constant.ServerResponseConstant;
 import com.staging.entity.Role;
+import com.staging.entity.RolePermission;
 import com.staging.entity.School;
 import com.staging.entity.SchoolType;
+import com.staging.service.RolePermissionService;
 import com.staging.service.RoleService;
 import io.swagger.annotations.ApiOperation;
+import org.hibernate.validator.constraints.SafeHtml;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,6 +21,9 @@ import org.springframework.boot.autoconfigure.AutoConfigureOrder;
 import org.springframework.web.bind.annotation.*;
 
 import org.springframework.stereotype.Controller;
+import org.w3c.dom.stylesheets.LinkStyle;
+
+import java.util.List;
 
 /**
  * <p>
@@ -36,6 +42,9 @@ public class RoleController {
     @Autowired
     private RoleService roleService;
 
+    @Autowired
+    private RolePermissionService rolePermissionService;
+
     /**
      * 分页查询
      * @param pagerLayui
@@ -53,6 +62,12 @@ public class RoleController {
         return p;
     }
 
+    @PostMapping("list")
+    @ResponseBody
+    public ServerResponse selectList(){
+        return ServerResponse.createBySuccess(roleService.selectList(null));
+    }
+
     /**
      * 添加角色
      * @param role
@@ -68,13 +83,30 @@ public class RoleController {
         return roleService.insert(role)?ServerResponse.createBySuccess(ServerResponseConstant.SERVERRESPONSE_SUCCESS_SAVE):ServerResponse.createByError(ServerResponseConstant.SERVERRESPONSE_ERROR_SAVE);
     }
 
+    /**
+     * 添加删除角色关联的权限
+     * @param role
+     * @return
+     */
     @PostMapping("delete")
     @ResponseBody
     public ServerResponse delete(Role role){
         if(role.getRole()=="超级管理员"){
             return ServerResponse.createByError("不可删除");
         }
-        return roleService.deleteById(role.getId())?ServerResponse.createBySuccess(ServerResponseConstant.SERVERRESPONSE_SUCCESS_DELET):ServerResponse.createByError(ServerResponseConstant.SERVERRESPONSE_ERROR_DELET);
+        //先删除角色下的权限
+        List<RolePermission> rolePermissions =  rolePermissionService.selectList(
+                new EntityWrapper<RolePermission>()
+                        .eq("rid",role.getId())
+        );
+        boolean result = rolePermissionService.delete(
+                new EntityWrapper<RolePermission>()
+                        .eq("rid",rolePermissions.get(0).getRid())
+        );
+        if(result){
+            return roleService.deleteById(role.getId())?ServerResponse.createBySuccess(ServerResponseConstant.SERVERRESPONSE_SUCCESS_DELET):ServerResponse.createByError(ServerResponseConstant.SERVERRESPONSE_ERROR_DELET);
+        }
+        return ServerResponse.createByError(ServerResponseConstant.SERVERRESPONSE_ERROR_DELET);
     }
 
     @GetMapping("page")
